requirements.txt
# Backend requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
firebase-admin==6.4.0
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Additional requirements
httpx==0.25.1
typing-extensions==4.8.0
